api/requirements/base.txt
Django==2.0.7
djangorestframework==3.8.2
django-extensions==2.0.7
psycopg2==2.7.5
django-cors-headers==2.3.0
django-dotenv==1.4.2
django-filter==1.1.0
django_crontab==0.7.1
django-redis==4.9.0
